# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Crowd-source questions for a meetup. Questioner helps the meetup organizer prioritize questions to be answered. Other users can vote on asked questions and they bubble to the top or bottom of the log",
    "version": "1.0.0",
    "title": "Questioner API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "sulaiman.sanusi@icloud.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "questioner-app-api.herokuapp.com",
  "basePath": "/api/v1",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/questions/{questionId}/downvote": {
      "patch": {
        "tags": [
          "Questions"
        ],
        "summary": "Downvote a Question by ID",
        "description": "",
        "operationId": "downVote",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation Or Meetup-Id Doesnt Exit"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication Errors"
          },
          "403": {
            "description": "Token Forbidden"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "security": [
          {
            "tokens": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "tokens": {
      "type": "apiKey",
      "name": "token",
      "in": "header"
    }
  },
...
}